ai/pharmguard/requirements.txt
# FastAPI and web server
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20

# HTTP client
httpx==0.28.1

# ChromaDB and embeddings (핵심)
chromadb==1.0.15
sentence-transformers==5.0.0

# Core ML dependencies (GLiNER가 필요로 함)
torch==2.6.0
transformers==4.53.3
tokenizers==0.21.2
huggingface-hub==0.33.4
safetensors==0.5.3

# GLiNER for Korean NER (핵심)
gliner==0.2.21

# Korean NLP (GLiNER-ko가 필요로 함)
konlpy==0.6.0
jpype1==1.6.0
python-mecab-ko==1.3.7
python-mecab-ko-dic==2.1.1.post2

# Google Cloud Vision API (핵심)
google-cloud-vision==3.10.2

# Image processing (OpenCV headless만 사용)
opencv-python-headless==4.10.0.84
Pillow==10.4.0

# Basic utilities
python-dotenv==1.1.1
requests==2.32.3
numpy==2.3.0

# Core Python dependencies
pydantic==2.11.7

# Data processing for vector DB ingestion
pandas==2.2.2